refactor: consolidate SRT parsing into single canonical implementation
Move SubtitleCue, _timestamp_to_seconds, and parse_srt_content from
sync.py into text_utils.py as the single authoritative SRT block-
splitting implementation. sync.py re-exports them for backward
compatibility. translation.py._parse_srt now delegates to the canonical
parse_srt_content() instead of reimplementing block splitting inline.
