+++ src/homelab_subs/core/sync.py
@@ -15,37 +15,25 @@ video release).
 
 from __future__ import annotations
 
-import re
 from dataclasses import dataclass
 from difflib import SequenceMatcher
 from pathlib import Path
 from typing import Callable, Optional
 
 from .audio import FFmpeg
-from .text_utils import normalize_text, format_timestamp
+from .text_utils import (
+    normalize_text,
+    format_timestamp,
+    SubtitleCue,
+    parse_srt_content,
+    _timestamp_to_seconds,
+)
 from .transcription import Segment, Transcriber, TranscriberConfig
 from ..logging_config import get_logger, log_stage
 
 logger = get_logger(__name__)
 
 ProgressCallback = Optional[Callable[[float, int], None]]
-
-
-@dataclass
-class SubtitleCue:
-    """
-    Represents a single subtitle cue with timing information.
-    """
-
-    index: int
-    start_seconds: float
-    end_seconds: float
-    text: str
-
-    @property
-    def duration(self) -> float:
-        """Duration of the cue in seconds."""
-        return self.end_seconds - self.start_seconds
 
 
 @dataclass
@@ -112,31 +100,6 @@ class SyncResult:
     max_offset_seconds: float
 
 
-def _timestamp_to_seconds(timestamp: str) -> float:
-    """
-    Convert SRT timestamp (HH:MM:SS,mmm) to seconds.
-
-    Parameters
-    ----------
-    timestamp : str
-        Timestamp in format "HH:MM:SS,mmm" or "HH:MM:SS.mmm"
-
-    Returns
-    -------
-    float
-        Time in seconds.
-    """
-    # Handle both comma and dot as millisecond separator
-    timestamp = timestamp.replace(",", ".")
-
-    match = re.match(r"(\d{2}):(\d{2}):(\d{2})\.(\d{3})", timestamp)
-    if not match:
-        raise ValueError(f"Invalid timestamp format: {timestamp}")
-
-    hours, minutes, seconds, ms = map(int, match.groups())
-    return hours * 3600 + minutes * 60 + seconds + ms / 1000
-
-
 # Backwards-compatible aliases for the canonical implementations in text_utils.
 _seconds_to_timestamp = format_timestamp
 _normalize_text = normalize_text
@@ -197,70 +160,8 @@ def parse_srt_file(srt_path: Path, encoding: str = "utf-8") -> list[SubtitleCue]
     return parse_srt_content(content)
 
 
-def parse_srt_content(content: str) -> list[SubtitleCue]:
-    """
-    Parse SRT content string into a list of SubtitleCue objects.
-
-    Parameters
-    ----------
-    content : str
-        SRT file content.
-
-    Returns
-    -------
-    list[SubtitleCue]
-        List of parsed subtitle cues.
-    """
-    cues: list[SubtitleCue] = []
-
-    # Normalize line endings
-    content = content.replace("\r\n", "\n")
-
-    # Split into blocks (separated by blank lines)
-    blocks = re.split(r"\n\n+", content.strip())
-
-    for block in blocks:
-        lines = block.strip().split("\n")
-        if len(lines) < 3:
-            continue
-
-        # Parse index
-        try:
-            index = int(lines[0].strip())
-        except ValueError:
-            logger.warning(f"Skipping invalid subtitle index: {lines[0]}")
-            continue
-
-        # Parse timestamp line
-        timestamp_match = re.match(
-            r"(\d{2}:\d{2}:\d{2}[,\.]\d{3})\s*-->\s*(\d{2}:\d{2}:\d{2}[,\.]\d{3})",
-            lines[1].strip(),
-        )
-        if not timestamp_match:
-            logger.warning(f"Skipping invalid timestamp: {lines[1]}")
-            continue
-
-        try:
-            start_seconds = _timestamp_to_seconds(timestamp_match.group(1))
-            end_seconds = _timestamp_to_seconds(timestamp_match.group(2))
-        except ValueError as e:
-            logger.warning(f"Skipping cue with invalid timestamp: {e}")
-            continue
-
-        # Remaining lines are the subtitle text
-        text = "\n".join(lines[2:]).strip()
-
-        cues.append(
-            SubtitleCue(
-                index=index,
-                start_seconds=start_seconds,
-                end_seconds=end_seconds,
-                text=text,
-            )
-        )
-
-    return cues
-
+# parse_srt_content is re-exported from text_utils for backward compatibility.
+# Callers that imported it from this module continue to work.
 
 def write_srt_from_cues(
     cues: list[SubtitleCue],
